refactor: validator 패키지를 application 패키지 안으로 이동
