+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -12,7 +12,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
-import kitchenpos.order.validator.OrderValidator;
+import kitchenpos.order.application.validator.OrderValidator;
 import org.springframework.data.annotation.CreatedDate;
 
 @Entity
