+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -18,8 +18,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.order.application.OrderService;
-import kitchenpos.order.validator.OrderValidator;
+import kitchenpos.order.application.validator.OrderValidator;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.order.domain.Order;
