+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -18,8 +18,7 @@ import static org.mockito.Mockito.doThrow;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.table.validator.TableValidator;
-import kitchenpos.table.application.TableService;
+import kitchenpos.table.application.validator.TableValidator;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.table.dto.response.OrderTableResponse;
